+++ CHANGELOG.md
@@ -40,3 +40,4 @@ Contributor to this version: Baptiste Hamon (@baptistehamon).
 ### Bug Fixes
 - Fix label error for projected suitability changes summary figures ([GH#11](https://github.com/baptistehamon/nzlusdb/issues/11), [PR#19](https://github.com/baptistehamon/nzlusdb/pull/19)).
 - Fix standardization function parameters error for apple chill units criteria ([GH#31](https://github.com/baptistehamon/nzlusdb/issues/11), [PR#33](https://github.com/baptistehamon/nzlusdb/pull/33)).
+- Fix cherry cracking survival indicator computation ([PR#36](https://github.com/baptistehamon/nzlusdb/pull/36))